MB-49934 Missing SORT operator for query with ORDER BY when HASH JOIN is used and building on outer side
If a query has ORDER BY clause, and a hash join is used, then the
probe side of the hash join can have its index order preserved.
Currently the decision of whether to preserve this.order (builder)
is made before the join itself it being planned, based on join
hint specified.

With CBO, when proper optimizer statistics is present, CBO can
switch the build/probe side of the hash join, when there is no
join hint specified. In such cases if we end up building on the
"outer" side of the join, then we are preserving index order on
the wrong side of the hash join. This can leads to wrong result
(missing SORT operator in some cases).

The fix is not to rely on join hint specified, rather, if we
decide to use hash join, we then check whether the hash join
is building on the outer side, and if it is, reset this.order
at that time.

Change-Id: Ie3593cf653ccfa17f4ff3856744d80a7488206fd
Reviewed-on: https://review.couchbase.org/c/query/+/167084
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>