MB-50041 Allow MISSING value for an alias from build side of HASH JOIN
A HASH JOIN can have on its build side another join, and if the inner
join is an OUTER join, it is possible to have MISSING value for one
of the aliases on the build side of the HASH JOIN. When processing
the HASH JOIN, current logic requires that all aliases from the build
side must have a valid value, this triggers an internal error for the
problematic query where the build side of the HASH JOIN is an another
OUTER join.

The fix is to simply skip the MISSING value for an alias instead of
returning an internal error in such cases.

Change-Id: I1c0f062357dd6d06578dbe42c07a3adc298fc2aa
Reviewed-on: https://review.couchbase.org/c/query/+/167433
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>