MB-51454 Allow Meta() functions to be used as null-rejecting expressions
For ANSI OUTER JOIN to ANSI INNER JOIN transformation, we check
for "null-rejecting" expressions specific to the subservient
side of the OUTER JOIN in the WHERE clause.

In 6.5.x, the process for checking "null-rejecting" expression
is flawed, as a result in the problematic query the expression
Meta().id LIKE <pattern> is treated as a null-rejecting
expression and thus the OUTER join is transformed into INNER
join. In 6.6.x and above, the process is fixed, however as a
result the expression involving Meta() function is no longer
treated as null-rejecting, and the OUTER join is NOT
transformed into an INNER join. Subsequently the WHERE clause
is not available when the join is being considered, resulting
in "no-index found" error.

The fix is to add a special case for Meta functions for
consideration of "null-rejecting" expressions.

Change-Id: I865ab72d3faad1a709ab00fe8d90a554506bf5ed
Reviewed-on: https://review.couchbase.org/c/query/+/167532
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/172921
Well-Formed: Restriction Checker