MB-49127 query infrastructure for jsevaluator usage in cloud
This change allows for javascript UDFs to be created using either libraries under specific scopes or under the base level (for on prem)
The idea is that if requests are tracked, then they are running in a cloud free tier environment, then we will pass to jsevaluator a library path (akin to query_context) whenever we execute a function.
The changes are not final - first I also need cover the case for non free cloud environments, and second the way jsevaluator gets a path from query needs to be finalized:
this implementation is what Jeelan, Abhishek and I have discussed this morning.

Change-Id: I34309fb4baa3e8cd5c022c12f4bca14fac3ab04a
Reviewed-on: https://review.couchbase.org/c/query/+/167772
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>