MB-50203 allow index scans under intersect scan to finish
Current implementation of intersect scan at execution time has a
mechanism for early-termination, i.e., as soon as one index scan
finishes a stop signal is sent to all other index scans, and the
result is based on the index scan(s) that has finished.

With CBO, the cost of an intersect scan is calculated assuming
all index scans finish, which provides a benefit in that some
index keys returned can be thrown out if an index key is not
common among all index scans. This reduces the number of
documents that needs to be fetched. However, with early
termination at execution time, this benefit does not always
materialize if some of the index scans do not finish and thus
cannot be used to throw away index keys.

We now provide a mechanism for all index scans under an
intersect scan to finish, such that the benefit can be fully
materialized. During execution, if all index scans are to
finish, we directly set childBits to fullBits, and avoid
sending a stop signal to the other index scans when the first
index scan finishes. Also there is no need to do incremental
update of childBits.

Change-Id: I8ec764b084a3dddeb7f7e34ba5bec6682692abc5
Reviewed-on: https://review.couchbase.org/c/query/+/168069
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>