MB-50300 Adjust cost for intersect scan
After the fix for MB-50203 intersect scan can execute faster for
certain queries, and is now causing bad "correlation" between
optimizer cost and execution time. The main reason is that the
cost of an intersect scan adds up the costs of all index scans
underneath, however at runtime all index scans are executing
in parallel and thus the cost is not accurately reflecting
the execution time.

We now take the max cost of the index scans underneath an
intersect scan to be the cost of the intersect scan, which
more accurately reflect runtime execution.

Change-Id: I128ec7a88f7c3ace017d8ceaf408920c6c6ef68e
Reviewed-on: https://review.couchbase.org/c/query/+/168347
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>