MB-50403 Proper calculation of hash join selectivity
If an index scan is used, the filters that are used to generate
the index spans should be marked with a special flag to indicate
that their selectivities are already accounted for.

For a keyspace on the inner side of an ANSI JOIN, since with CBO
both nested-loop join and hash join are considered, previously
the special index flag is only (permenantly) marked after the
join method is chosen. However this causes issues since when
cost is calculated for a nested-loop join or a hash join, we
also need to exclude the selectivities for the filters marked
with the special index flag. For nested-loop join this is working
properly since we temporarily mark the filters before the cost
is calculated; for hash join this is not working properly, due to
lack of proper marking of this special index flag.

The fix is to change the way we mark this special index flag.
For inner side of ANSI JOIN, when the inner side is being planned
for either nested-loop join or hash join, also mark the special
index flag, and make sure this marking is cleared when a join is
being considered. The permenant marking is done after we've
chosen the join method.

Change-Id: I807d779bd9a5e630065a7fbe5f7180c3cbce63e1
Reviewed-on: https://review.couchbase.org/c/query/+/168729
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>