MB-50417 Save/Restore index flag during join enumeration
Continuation from MB-50403, although the issue exists before that as well.

The special index flag can potentially be reset every time the inner
keyspace is replanned. Without join enumeration this is ok since
each keyspace is planned once, with both nested-loop join and hash
join considered, either will replan the inner keyspace. However with
join enumeration this is no longer the case, since each keyspace can
be replanned many times, and also when hash join is being considered
we no longer replan the inner keyspace, we just use a previously
generated plan. Thus in case of join enumeration we could
potentially mess up the index flag on the inner keyspace filters,
resulting in errors in cost calculations.

The fix is to save/restore the index flag on each filter when a
nested-loop join is being considered during join enumeration. This
is not needed for hash join since we don't replan the inner keyspace
for hash join.

Change-Id: Iaa8931f7338617595bfed214a1783b750dfdd7d7
Reviewed-on: https://review.couchbase.org/c/query/+/168786
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>