MB-50418 Avoid processing ON-clause twice for outer join during join enumeration
The ON-clause of an outer join cannot be pre-processed. Previously
we process the ON-clause when the join is being considered. With
join enumeration, we also need to pre-process the ON-clause when
single-keyspace plan is being considered. This currently results
in the ON-clause of an outer join being processed twice, i.e.,
each individual filter in the ON-clause is added to the filters
array of baseKeyspace twice.

The fix is to avoid processing the ON-clause for outer join during
join enumeration. In addition, the logic for determining whether
an ON-clause is "pushable" for a non-outer join is also updated.

Change-Id: I1ff4bf36114a17371edcb364e8377e7a81b44548
Reviewed-on: https://review.couchbase.org/c/query/+/168787
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>