MB-50541 Properly update index statistics for primary index
With INDEX ALL option we gather all indexes defined
and gather a list of terms to be used for UPDATE
STATISTICS. Current code explicitly excludes a primary
index since it does not have any terms. However this
results in index statistics for a primary index not
properly updated.

The fix is to leave the primary index in the list of
indexes, it'll not contribute to any terms but its
index statistics still should be updated.

Change-Id: Id5527df1d373a251557466538e7ed47f24744bc9
Reviewed-on: https://review.couchbase.org/c/query/+/169276
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>