MB-50724 Possibly return reason in 13014
We have an inconsistency in that depending on how crendtials are
provided the error for failed authentication differs.

In order to not break existing usage, this fix augments 13014
(authorisation failure) with the authentication error (if there
is one) *only when impersonation is taking place* so that a clear
distinction between authentication and authorisation can be made.

Limiting to impersonation only reduces scope to where it is really
needed - for other services that need to distinguish between
authentication and authorisation failures.

Change-Id: I1d26ed0572e39d36e582c4fcb2f1fadc82fd7aec
Reviewed-on: https://review.couchbase.org/c/query/+/169480
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>