MB-50649 Allow constant array inside IN-clause to use set semantics
Currently in a constant array construct, the position of elements
inside the array is significant, e.g. when comparing two arrays
the element at the same position is compared.

For an IN-clause, the elements of the array at the right-hand side
is really a set, i.e. the positions of the elements are not
significant in the evaluation of the IN-clause expression. E.g.
"c1 IN [1,2,3]" is functionally equivalent to "c1 IN [2,3,1]".

With CBO, we need to match predicates constructed from the index
spans with the original predicates. When an IN-clause is used, for
EXPLAIN stability we sort the elements of the IN-list before
generating the index spans. This means that the constructed
elements in the IN list may differ from the original IN list.

Previous attempt to fix this handles a simple IN predicate but
does not handle more complicated situations like when IN clause
is used within an ANY clause.

We now use an indicator inside ArrayConstruct to indicate whether
set semantics should be used. When an IN expression is constructed,
if the right-hand side is an array construct, set it to use set
semantics.

Change-Id: If1cf021b2d69fba6e8233a3378959a248f2edbb4
Reviewed-on: https://review.couchbase.org/c/query/+/169624
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>