MB-50652 Create N1QL_SYSTEM_BUCKET outside of dictionary lock
When the system bucket N1QL_SYSTEM_BUCKET is being created,
set the status to _CREATING (already there), and use a new
Condition variable for situations where the caller need to
wait for the CREATE operation to complete. For other
situations where waiting is not necessary, just return as if
the system bucket does not exist. When the create completes
broadcast to all waiters of the Condition variable.

Change-Id: I01859f138ab3eafb55a26a36e8325462b07ece4a
Reviewed-on: https://review.couchbase.org/c/query/+/169691
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>