MB-50716 Reuse prepared admin operations where possible.
When processing a PreppareTopologyChange callback, we previously
regenerated the prepared admin operations for all nodes in the KeepNodes
list before assigning this as the active servers list.  This change aims
to reduce the processing time in PrepareTopologyChange by reusing any
prepared operations for nodes in the existing servers list.  In the
majority of cases reuse makes good sense but it does mean that, should it
be possible for the underying query node information to change, the
prepared operation will be stale and never be replaced with an updated
version.

This also caches the host name so as to avoid repeated UUIDToHost
calls, again primarily to save processing time in PrepareTopologyChange
invocations.

Correct parallel node ejection issue.

Change-Id: I334c0be78cb42f29554ecb533a021cb649806cb7
Reviewed-on: https://review.couchbase.org/c/query/+/170091
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>