MB-50735 restrict javascript library usage in function creation
the create function ... language javascript  ... at clause now inspects the library name.

Only the following is allowed:

<library> this means a root library name for non tenant environments, and /bucket/scope/library for tenant environments
./<library> this means library for global libraries and /bucket/scope/library for scope libraries (tenant and not)
/<bucket>/<scope>/<library> for scope libraries, where the only bucket and scope that can be specified is the bucket and scope of the function.

I expect that ./library would be the standard usage, as it is portable across all scenarios.
In the future we might allow cross scope references - although in order to do so, jsevaluator has to be extended to authorize users.

Change-Id: I4967763692f6577bd2f661cc9cc07d51eee4d19b
Reviewed-on: https://review.couchbase.org/c/query/+/170178
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>