MB-21959 Implement switchable duration output format
Primarily a request parameter (duration_style) but a default setting
(duration-style) is also provided but is not persisted by ns_server.

The default can only be set via the REST interface and only permits
setting to LEGACY or COMPATIBLE so that the UI is unaffected.

Attempts are made to understand all duration formats when parsing
regardless of duration style setting, except for str_to_duration when
an explicit style is provided as the second argument.  (This is an
effort to ease use.)

Fractions for non-legacy formats have trailing zeros trimmed to smallest
necessary unit grouping - i.e. ms (3 digits), µs (6 digits) or ns (9
digits).

Example output:

  cbq> \set -duration_style legacy;    # or don't set it
  cbq> select uptime from system:vitals v;
  {"uptime":"35m55.077183175s"}

  cbq> \set -duration_style seconds;
  cbq> select uptime from system:vitals v;
  {"uptime":"2160.856697570"}

  cbq> \set -duration_style compatible;
  cbq> select uptime from system:vitals v;
  {"uptime":"2166.406815924s"}

  cbq> \set -duration_style interval;
  cbq> select uptime from system:vitals v;
  {"uptime":"0:36:12.435039902"}

Change-Id: I158e7212b87ae6068c063380a7d20f41d0b4c662
Reviewed-on: https://review.couchbase.org/c/query/+/190100
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>