MB-51457 Check collection Uid before reusing indexer objects
When changes were made to collections outside of query we could miss
that the collection had been dropped and a new collection created with
the same name and hence were incorrectly reusing the indexer objects.

With this change we confirm the UID before reusing the objects when
refreshing the collections.

Change-Id: Ia1ffb0d1087c7aa4fd326ad920faa6b27c6c9614
Reviewed-on: https://review.couchbase.org/c/query/+/172608
Well-Formed: Restriction Checker
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>