MB-51529 Proper handling of filters that does not reference a keyspace, but is not "constant"
If the WHERE clause contains a "constant" filter, e.g.
    WHERE true
or
    WHERE 1 = 0
the planner is able to detect and mark such constant filters, however,
if a filter does not reference any keyspace, but is not "constant",
e.g. the filter in this defect:
    WHERE floor(random()*10) = 5
currently the planner simply ignore the filter when it classifies
filters and attach to keyspaces.

The fix is to detect and mark such situations such that when index
pushdown is considered we set "exact" to FALSE if such a filter
exists.

Change-Id: I427e69771212afd2944239efedd16c318e1f779b
Reviewed-on: https://review.couchbase.org/c/query/+/172620
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>