MB-51561 Proper handling of empty span at runtime in case of multiple spans
At runtime, we evaluate index span expressions and generate the
actual index spans for the indexer. In case of multiple spans,
current code incorrectly inherit the "empty" indicator for the
last index span evaluated, i.e., if the last index span is
empty, then current code returns true for the empty indicator,
resulting in index scan not being done.

There is a separate issue with planner considering OR-clauses
and an error 4321 is returned if an arm of the OR only
references named/positional parameters.

The fix has three parts:

1. when evaluating a single index span, if one of the ranges
   is empty, then the entire span is empty, there is no need
   to add the span. E.g., index keys on (c1, c2), and c2 range
   is empty, then there is no need to add an index span for
   just c1.
2. set the return empty indicator based on whether any index
   span generated.

3. fix the logic for considering OR-clauses in planner when
   an arm of the OR only references query parameters

Change-Id: Iea0eb1cb88d5ac822ea9133b82a6184c3c949897
Reviewed-on: https://review.couchbase.org/c/query/+/172755
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>