MB-15940 Add VALIDATE to USE KEYS clause
The new syntax is USE KEYS [VALIDATE] ...  / ON KEY|KEYS [VALIDATE] ...

This is intended to assist with diagnosing query problems where a USE
KEYS clause is in use.  With the keyword present, missing keys are
reported as code 5503 warning message.

A single warning is issued for each keyspace and lists the total number
of invalid/missing keys along with up to 10 keys.  If a key is repeated
in the clause and is in the first 10, it is repeated in the warning list
to highlight the duplication.

Limitations:
- If no items are produced by the outer leg of a join, validation on
  the inner leg does not take place.
- If there are no items in the keyspace at all, no validation takes
  place.

Change-Id: I46ec70d56d7f711c8415d9a02792941d1660b2e9
Reviewed-on: https://review.couchbase.org/c/query/+/172849
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>