MB-51669 Make sure UnderNL flag is unset with CBO
When considering joins the planner sets UnderNL flag on the
inner keyspace during consideration of nested-loop join, when
scan planning is being done on the inner keyspace. After that,
it should unset the UnderNL flag unless the planner chooses
NL join.

This is not an issue when RBO is being used, since we only
consider one join method with RBO. With CBO, in absence of
join hint, and when optimizer stats are available, both
NL join and hash join are considered. In 7.0, we did not
properly unset the UnderNL flag when a hash join is chosen.
This is fixed in 7.1, however, the fix is incomplete in
that the UnderNL flag is not unset when an error is returned,
e.g. in case of missing secondary index on the inner.
With join enumeration, multiple plans are considered and
it's possible that one join consideration results in such
an error, and then incorrectly carry the UnderNL flag.

The fix is to unset the UnderNL flag immediately after scan
planning of the inner keyspace, and only reset the UnderNL
flag if a NL plan is chosen.

Change-Id: I13f8911a13819c346c22d2f4982a38f8d525f92f
Reviewed-on: https://review.couchbase.org/c/query/+/173174
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>