MB-51701 Proper handle join hint on first keyspace
When join enumeration is not done, e.g. lacking necessary
optimizer statistics, or when ORDERED hint is used, then
current code is not properly handling join hint specified
on the first keyspace. (Note we don't allow legacy join
hint to be specified on the first keyspace, however this
restriction does not apply to the new relational-style
join hint).

Make sure when we process the first keyspace (without join
enumeration) we mark the join hints as not followed.

Also fixed a bug such that in case of right join or merge,
we only mark the INFER/TRANSFER property when there is
actually a legacy-style join hint specified.

Change-Id: If592803d5e92011293fdbcaff5b6e60f8109cf9a
Reviewed-on: https://review.couchbase.org/c/query/+/173266
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>