MB-51741 Additional changes for IN-list handling
The hash table usage for IN-list evaluation requires explicit
setup/teardown on the expression itself. This is already done
for the expression in Filter operator, as well as early filters
for the join/nest operators. However, there are other filters
that currently do not have this setup/teardown and thus the
hash table optimization is not kicking in.

Add the necessary setup/teardown logic for other places where
early filters are evaluated at runtime, including:
  - covering index scan
  - expression scan
  - unnest

Change-Id: I408628728a9151a56a221b1df8f59f0d088703ea
Reviewed-on: https://review.couchbase.org/c/query/+/173312
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/173313