MB-51753 Detect and mark CTE variable as correlated
The binding expression for a CTE could be correlated if the
CTE is inside a subquery. Current correlation is not detected
which could result in the subbquery itself not properly marked
as correlated.

Add logic to detect and set correlated variable identifiers,
and set a subquery to be correlated if such correlated
variable exists.

In addition, if the CTE variable is used on the right-hand
side of an IN-list, avoid using hash table optimization if
the CTE variable is correlated.

Change-Id: Ia13b591ed770a473026c045161a9097faaa87a43
Reviewed-on: https://review.couchbase.org/c/query/+/173381
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>