MB-51220 Remove constant expressions from Filter evaluation
Implement an expression mapper that checks and removes constant
expressions from an expression tree. It removes TRUE subterms
under AND, and FALSE subterms under OR. If the entire expression
tree is constant, it returns either a TRUE_EXPR or a FALSE_EXPR.

This is then used to remove constant expressions from the WHERE
clause.

Change-Id: I42021e44d6a8cc2b0350cd7063d5ce6bd839b999
Reviewed-on: https://review.couchbase.org/c/query/+/173444
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>