MB-51739 Properly handle derived filters in post-join filter determination
The issue with this defect is there is a BETWEEN filter in the
ON-clause of an outer join. BETWEEN filter is transformed into
a pair of GE and LE filters. When the join is generating
early filters it incorrectly determined that the LE part of
the filter is post-join, and applied it at the join which is
incorrect. The filter is part of ON-clause and not post-join.

The fix has two parts:

1. When BETWEEN filter is transformed into GE/LE pair of filters
   mark the generated filters with DERIVED flag
2. During determination of post-join filters, if the filter
   is derived, also transform the ON-clause filter in order to
   determine SubsetOf.

Change-Id: Ib6fd144cab6fabf50e37b9a032fb2d496e8e4b5b
Reviewed-on: https://review.couchbase.org/c/query/+/173310
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/173594
Well-Formed: Restriction Checker