MB-51816 Proper marking of all filters used in Index Scan
With CBO, when an index scan is generated, we mark the keyspace
filters used in that index scan such that the selectivity of those
filters are not reapplied (already applied at the index scan).
This process currently takes care of all index spans and index
condition, but it does not process potential index scan filters
(early filters put on covering index scan).

Change-Id: I0f57906544be3ef67207ce90808fab7815b99d2f
Reviewed-on: https://review.couchbase.org/c/query/+/173657
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>