MB-52061 Check hash join hint on both sides of join during join enumeration
With introduction of negative join hints the way we handle join hint
is changed. We no longer "transfer" join hint during join enumeration,
instead we should be checking for join hint on both sides, if left-hand
side is also a SimpleFromTerm. This is currently not done properly.

The fix is to:
  1. set left-hand side of join node if it is SimpleFromTerm
  2. check for join hint from left-hand side during join enumeration
     if join hint is not already specified on right-hand side
  3. handle negative join hints in compatibleJoinHints()

Change-Id: Ib8870ae261f214cb830f7ba64eda7fec2c943820
Reviewed-on: https://review.couchbase.org/c/query/+/174556
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>