MB-51928 Costing changes for join filter
With join enumeration:

A new step is added when constructing single keyspace
plans, to also consider potential join filters pushed
to this keyspace. Such plans are marked and tracked
during join enumeration as the "source" of join filters
are being joined in. Size of each join filter is
calculated based on optimizer statistics.

Without join enumeration but with CBO:

As plan are being constructed consider join filters
when appropriate. Size of each join filter is
calculated based on optimizer statistics.

Without CBO:

Join filter is only considered with JOIN_FILTER hint.
The size of each join filter depends on the total
document count of the source keyspace. If that is not
available, use a default.

Change-Id: Ifac57a944eb3d81a9f1212bb38a2794d0e301d4a
Reviewed-on: https://review.couchbase.org/c/query/+/174617
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>