MB-51817 Reset covers on annotatedValue for Fetch
In case we use index filters or join filters (bloom filters)
on index keys, the annotatedValue used will have "covers"
information to keep track of the values for the index keys.
When we do Fetch, remove these covers since they will no
longer be needed.

This is done unconditionally since it's a rather trivial
operation. The alternative is to add an indicator on the
Fetch operator with an extra if check, which seems
rather unnecessary for such a simple task.

Change-Id: Id9bdf3c2a2010bba19e979429171d1ca9555d150
Reviewed-on: https://review.couchbase.org/c/query/+/174689
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>