MB-51817 Include fetch cost when comparing indexes under CBO
Previously under CBO when we compare two indexes we use the
cost of index scans, this imiplicitly assumes that the
two index scans produces the same number of keys to feed
into fetch and thus the cost of fetch is the similar
(which is not true even before).

With index filters, it is possible for different index
scans to apply a different set of filters on the index,
and thus the number of resulting keys may be different, as
a result the cost of the subsequent fetch may also be
different. In order to properly compare two index scans,
keep track of fetch cost in index entry, and include fetch
cost when comparing costs of two index scans.

Change-Id: Ic0783b744b92cbf54a3c672a3e2eb726ae8fb9ea
Reviewed-on: https://review.couchbase.org/c/query/+/174736
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>