MB-52090 Compare with original index condition for index eligibility
When the index condition has LIKE predicate, when we formalize the
expression we also perform DNF transformation. This is normally not
an issue since the query predicate would also go through DNF
transformation as well. In case of this bug, the query predicate
uses a query parameter, and we did not perform DNF transformation
before hand due to the presence of query parameter. As a result,
when we match the query predicate with index condition to determine
index eligibility, we did not find a match and discarded the index.

The fix is to also compare with the original index condition
(before DNF transformation) as well for determining index eligibility.

Change-Id: I728da90f97b1159f0833cd88a43a60f17bb152d8
Reviewed-on: https://review.couchbase.org/c/query/+/174778
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>