MB-51817 Handle OR predicate better
Currently implementation of index filters relies on marking each
individual filter as "exact", such that we don't need to reapply
exact filters on the index keys.

The repro for CBSE-8622 has an OR predicate (derived from != )
and in this case, currently when we do sargIndexes to generate
index spans we don't use individual filters, and thus we don't
mark the filters as exact, resulting in filters being reapplied
unnecessarily.

For OR predicates we go through two passes, one when we
consider the entire OR clause, and one when we use individual
subterms of the original OR predicate. Mark the case where we
use the entire OR clause and in such case allow using
individual filters to sargIndexes. This allows the individual
filters to be marked as exact.

As a side note, there is currently no CBO based comparison
of index paths when we compare potential index paths from
the two passes mentioned above. Also added CBO based picking
or the right index path according to cost.

Change-Id: I3e5cc16062b5355371e8a0d7079d241c9834bf8d
Reviewed-on: https://review.couchbase.org/c/query/+/174783
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>