MB-52137 accept two part query contexts
this change now allows query contexts of the form namespace:bucket and
relative keyspace paths of the form scope.collection.

all the previous behaviour is supported, and a query of the form

select * from keyspace

with a query context of the form

namespace:bucket

will silently use namespace:bucket._default.keyspace for backwards compatibility.
conversely

select * from scope.keyspace

with a query context of the form

namespace:bucket.scope

will throw an error.
Ambiguities in grant and revoke are discussed in the comments.

Change-Id: I25d5ca3fe74a62251cc3a800727cb3512a582ddb
Reviewed-on: https://review.couchbase.org/c/query/+/174801
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>