MB-52155 Avoid adding duplicated index filter covers internally
For index filter covers, we used to use map[string]value.Value
to represent filters in index conditions. The fix for MB-49859
changed that to map[expression.Expression]value.Value, such that
the expression flags previously set are preserved. However, one
side effect of that is, when we generate filter covers from
index conditions, we use both the original condition as well as
potentially transformed condition to generate filter cover, and
in case of a simple expression, e.g. type = "airport", the
original condition and transformed condition are the same
expression, although the actual expression pointers are different.
Previously since we use the expression string as the key to the
map, duplicated expressions are only added once. Now we use the
expression pointer as the key to the map, the same expression
(duplicated, but with different expression pointers) are added
twice in filter covers.

The fix is to use an auxillary map with string as key to help
eliminate duplicates in function mapFilterCovers().

Change-Id: Ibc7a71f21ace8050240104257b921d96336dbc9d
Reviewed-on: https://review.couchbase.org/c/query/+/174817
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>