MB-52260 Properly identify keyspace references for Elements
The planner classifies expressions based on keyspace references
in the expression. Currently for a Field expression it only
checks for the First part (left-hand side of the Field), which
implicitly assumes the Second part is FieldName. However there
are situations where the second part can also reference a
keyspace alias, such as when it contains an Element reference.

The fix is to also check the second part of a Field expression
if it is not a FieldName.

Change-Id: Ie44dce9a55bd0b5b548356003bc9bedd8b9e0788
Reviewed-on: https://review.couchbase.org/c/query/+/175224
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/175233
Well-Formed: Restriction Checker