MB-52254 Perform SORT early when possible to save on FETCH
When a query has ORDER BY with LIMIT/OFFSET, if the query is not
covered and thus the ORDER/LIMIT/OFFSET cannot be pushed to the
index scan, currently we need to fetch all documents from the
index scan, perform SORT with LIMIT. This could be inefficient
if a large number of documents need to be fetched and the number
of documents returned is small due to LIMIT.

There are situations where the SORT with LIMIT can be performed
early if all the expressions in the SORT is available from the
index scan. In such situations it is beneficial to do the SORT
early and reduce the number of fetches that needs to be done.

1. When considering an index, check whether the ORDER/LIMIT
  info is available, i.e. simple query with ORDER/LIMIT, the
  ORDER/LIMIT info on builder will be reset for more complex
  queries such as presence of joins/nests, group by/aggregates
  etc.
2. Both ORDER and LIMIT are required to consider this optimization
3. There could be potential projection alias references in the
  SORT expressions, check and replace any such projection aliases
  with the actual expression.
4. Check if all the SORT expressions are available from the index
  (index keys and index conditions).
5. When comparing index scans, prefer indexes with early sort
  after other comparison criteria fails.
6. With CBO, recalculate the cost of the Fetch with reduced number
  of documents.
7. After building index scan, check whether early sort should be
  done, and build ORDER/LIMIT operator if needed. Mark that this
  is done.
8. At the end of the plan, when ORDER is normally added, check
  whether SORT is already done early, and skip ORDER if already
  done.
9. Index advisor check and retain recommendations with early order.

Change-Id: Iab5c79229b7f20f0cf203c8574f22a30f20566b4
Reviewed-on: https://review.couchbase.org/c/query/+/175306
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>