MB-52254 Perform OFFSET early when possible
With early order, the OFFSET is not applied at the Order
operator, instead, it is applied at the end. This is
reasonable if there exists predicates that's not applied
at the index as "exact" spans. However, in case the
index scan applies all the predicates, then it's feasible
to apply the OFFSET after the early order, to further
reduce the documents that need to be fetched.

Similar approach can be done in case the index provides
ordering but Fetch is still required (and OFFSET is not
already pushed to the index scan).

Change-Id: Ibbd7379e5abcb2cce23edf820e35913199685509
Reviewed-on: https://review.couchbase.org/c/query/+/175403
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>