MB-31640 Include subqueries in FROM clause and setop arms
For SubqueryTerm, its query plan is included in the main query
plan, however if the SubqueryTerm itself contains additional
nested subqueries, previous changes did not include such nested
subqueries in the explain.

The fix is to include the nested subqueries when planning
subqueries. Also moved the planning of subqueries to the
beginning of a statement instead of the end, this should help
CBO since the cost information from the subquery can be then
used in the main query. Also need to pass through the new
builder flag when planning subselects.

Change-Id: Ibb9058ccc006bf0851765a30b9c742b55fcdd69a
Reviewed-on: https://review.couchbase.org/c/query/+/175484
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>